requirements.txt
# Core dependencies
watchdog>=3.0.0
PyYAML>=6.0
requests>=2.28.0
rarfile>=4.0
tqdm>=4.64.0

# Additional dependencies for enhanced setup
lxml>=4.9.0
psutil>=5.9.0

# Optional GUI dependencies
pystray>=0.19.0
Pillow>=9.0.0

# Database
# sqlite3 is included in Python standard library

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0 